Produce a wheel for python 3.12
- test now on 3.9 and 3.12
- beware to some dependencies conflicts that make
   - install pygame 2.1.3 not compatible with python 3.12 (=> we enforce
     pygame >=2.5 in tests
   - install cffi 1.15.0 instead of more recent versions that are
     adapted for poython 3.12 (=> we enforce cffi>=1.17 in tests)
- up-tamer depends on pytamer not available for macos x86_64 + python >
  3.11 (was already not available on macos amd64)
- random.randint() must use real integers

We also update minizinc binary used in tests to 2.8.5.

+++ skdecide/hub/solver/cgp/pycgp/cgp.py
@@ -176,14 +176,16 @@ class CGP:
                         # mutate connection
                         self.genome[index] = rnd.randint(
                             0,
-                            min(
-                                self.max_graph_length + self.num_inputs - 1,
-                                (
-                                    self.num_inputs
-                                    + (int(index / (self.max_arity + 1)) - 1)
-                                    * self.num_rows
+                            int(
+                                min(
+                                    self.max_graph_length + self.num_inputs - 1,
+                                    (
+                                        self.num_inputs
+                                        + (int(index / (self.max_arity + 1)) - 1)
+                                        * self.num_rows
+                                    )
+                                    * self.recurrency_distance,
                                 )
-                                * self.recurrency_distance,
                             ),
                         )
                         # self.genome[index] = rnd.randint(0, self.num_inputs + (int(index / (self.max_arity + 1)) - 1) * self.num_rows)
